requirements.txt
colorama==0.4.4
pyspellchecker==0.6.3
